refactor: minor file + code organization fixes
