+++ src/components/About/OurAlumnis.tsx
@@ -2,35 +2,24 @@ export default function OurAlumnisPage() {
   return (
     <div className="relative flex flex-col gap-5 bg-[#8A8A8A] px-60 py-28">
       <div className="relative flex flex-col gap-14">
-        <h1 className="relative font-heading text-3xl font-bold text-white">
-          Our Alumnis
-        </h1>
+        <h1 className="relative font-heading text-3xl font-bold text-white">Our Alumnis</h1>
         <p className="relative flex items-center self-stretch text-xl font-normal text-white">
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit. Lorem ipsum
-          dolor sit amet, consectetur adipiscing elit. Lorem ipsum dolor sit
-          amet, consectetur adipiscing elit.
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit. Lorem ipsum dolor sit amet, consectetur adipiscing
+          elit. Lorem ipsum dolor sit amet, consectetur adipiscing elit.
         </p>
         <div className="relative flex flex-col gap-12 self-stretch">
           <div className="relative flex flex-col gap-5 self-stretch">
             <div className="relative flex flex-col gap-3">
-              <div className="relative w-32 font-heading text-2xl text-white">
-                2022
-              </div>
+              <div className="relative w-32 font-heading text-2xl text-white">2022</div>
               <div className="visible absolute left-20 top-2">
-                <svg
-                  width="17"
-                  height="9"
-                  viewBox="0 0 17 9"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
+                <svg width="17" height="9" viewBox="0 0 17 9" fill="none" xmlns="http://www.w3.org/2000/svg">
                   <line
                     y1="-0.75"
                     x2="12.3794"
                     y2="-0.75"
                     transform="matrix(0.686624 0.727013 -0.607597 0.794245 0 0)"
                     stroke="white"
-                    stroke-width="1.5"
+                    strokeWidth="1.5"
                   />
                   <line
                     y1="-0.75"
@@ -38,23 +27,14 @@ export default function OurAlumnisPage() {
                     y2="-0.75"
                     transform="matrix(-0.686624 0.727013 0.607597 0.794245 17 0)"
                     stroke="white"
-                    stroke-width="1.5"
+                    strokeWidth="1.5"
                   />
                 </svg>
               </div>
             </div>
             <div className="visible relative shrink-0 translate-y-1 self-stretch overflow-visible rounded-none">
-              <svg
-                width="967"
-                viewBox="0 -2 967 10"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-              >
-                <path
-                  d="M967 0L2.98023e-06 0"
-                  stroke="white"
-                  stroke-width="3"
-                />
+              <svg width="967" viewBox="0 -2 967 10" fill="none" xmlns="http://www.w3.org/2000/svg">
+                <path d="M967 0L2.98023e-06 0" stroke="white" strokeWidth="3" />
               </svg>
             </div>
           </div>
